Factor out classes from main file
