+++ ESP8266/Timer.py
@@ -1,0 +1,5 @@
+class Timer:
+  def __init__(self):
+    self.millis = 0
+    self.tick = 0
+    self.cur_stop = 0
